requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
APScheduler==3.11.2
attrs==26.1.0
babel==2.18.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
claude-agent-sdk==0.1.50
click==8.3.1
courlan==1.3.2
cryptography==46.0.5
curl_cffi==0.14.0
dateparser==1.3.0
defusedxml==0.7.1
Deprecated==1.3.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.2
feedparser==6.0.12
frozenlist==1.8.0
google-api-core==2.30.0
google-auth==2.49.1
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
htmldate==1.9.4
http_ece==1.2.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jusText==3.0.2
lingua==4.16.2
lingua-language-detector==2.1.1
lxml==6.0.2
lxml_html_clean==0.4.4
mcp==1.26.0
mss==10.1.0
multidict==6.7.1
nodriver==0.48.1
numpy==2.4.4
openai==2.29.0
polib==1.2.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.6
py-vapid==1.9.4
py3langid==0.3.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
pywebpush==2.3.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rpds-py==0.30.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
sse-starlette==3.3.3
starlette==1.0.0
tavily-python==0.7.23
tiktoken==0.12.0
tld==0.13.2
tqdm==4.67.3
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.42.0
websockets==16.0
wrapt==2.1.2
yarl==1.23.0
youtube-scraper==1.0.3
youtube-transcript-api==1.2.4
